secure creatives module, domain macro comment, review notes
