+++ src/prebid.js
@@ -6,6 +6,7 @@ import { videoAdUnit, hasNonVideoBidder } from './video';
 import 'polyfill';
 import {parse as parseURL, format as formatURL} from './url';
 import {isValidePriceConfig} from './cpmBucketManager';
+import {listenMessagesFromCreative} from './secure-creatives';
 
 var $$PREBID_GLOBAL$$ = getGlobal();
 var CONSTANTS = require('./constants.json');
@@ -289,58 +290,6 @@ function setRenderSize(doc, width, height) {
     doc.defaultView.frameElement.width = width;
     doc.defaultView.frameElement.height = height;
   }
-}
-
-function listenMessagesFromCreative() {
-  addEventListener('message', receiveMessage, false);
-}
-
-function receiveMessage(ev) {
-  var key = ev.message ? 'message' : 'data';
-  var data = {};
-  try {
-    data = JSON.parse(ev[key]);
-  } catch (e) {
-    return;
-  }
-
-  if (data.adId) {
-    const adObject = $$PREBID_GLOBAL$$._bidsReceived.find(function (bid) {
-      return bid.adId === data.adId;
-    });
-
-    if (data.message === 'Prebid Request') {
-      sendAdToCreative(adObject, data.adServerDomain, ev.source);
-      events.emit(BID_WON, adObject);
-    }
-  }
-}
-
-function sendAdToCreative(adObject, remoteDomain, source) {
-  const { adId, ad, adUrl, width, height } = adObject;
-
-  if (adId) {
-    resizeRemoteCreative(adObject);
-    source.postMessage(JSON.stringify({
-      message: 'Prebid Response',
-      ad,
-      adUrl,
-      width,
-      height
-    }), remoteDomain);
-  }
-}
-
-function resizeRemoteCreative(adObject) {
-  const { adUnitCode, width, height } = adObject;
-  const iframe = document.getElementById(window.googletag.pubads()
-    .getSlots().find(slot => {
-      return slot.getAdUnitPath() === adUnitCode ||
-        slot.getSlotElementId() === adUnitCode;
-    }).getSlotElementId()).querySelector('iframe');
-
-  iframe.width = '' + width;
-  iframe.height = '' + height;
 }
 
 //////////////////////////////////
